+++ changelog.md
@@ -1,8 +1,8 @@
 # Version 2025.1.12 (2025-01-23)
 
-- Refactor parameter_visibility
 - Add LOW_BAT_LIMIT
 - Add calculated data points: OperatingVoltageLevel
+- Refactor parameter_visibility
 
 # Version 2025.1.11 (2025-01-20)